Make Client and Websocket classes customizable
If you want to use a non-standard way to login into Mattermost, you need to subclass the `Client` and `Websocket`classes (and maybe even the `Driver`). 

However, `Driver` will not use your custom subclasses. A clean way to inject them into the Mattermost `Driver` would be to just pass them to `Driver()` and `Driver.init_websocket()`.