Add missing status, roles and commands properties
While attempting to modify a user's status I noticed the property for accessing the endpoint didn't exist on the `Driver` class but did exist under the deprecated api property. I'm also adding the roles/commands properties too while I'm here so the properties the `Driver` class should now be in sync with what's available in the deprecated api property.